move out most of AssemblerConnectionManager into DatabaseQueryService, a couple other minor fixes
